Refactor: Move subculture classes into subculture/ module
